chore(plugin): upgrade to RuneLite 1.12.34.1 and drop deprecated API usage
Scene, coords and pathfinding APIs read through the top-level WorldView;
prayer state reads each prayer's varbit; projectile events use
getSourcePoint/getTargetActor (src_x/src_y are now world coordinates —
noted in BASE.md); graphic events iterate spot anims; inventory uses
gameval InventoryID.

+++ plugin/src/main/java/dev/ragger/plugin/scripting/CameraApi.java
@@ -36,9 +36,9 @@ public class CameraApi {
     public double focal_x() { return client.getCameraFocalPointX(); }
     public double focal_y() { return client.getCameraFocalPointY(); }
     public double focal_z() { return client.getCameraFocalPointZ(); }
-    public void set_focal_x(final double x) { client.setCameraFocalPointX(x); }
-    public void set_focal_y(final double y) { client.setCameraFocalPointY(y); }
-    public void set_focal_z(final double z) { client.setCameraFocalPointZ(z); }
+    public void set_focal_x(final double x) { client.setCameraFocalPointX((float) x); }
+    public void set_focal_y(final double y) { client.setCameraFocalPointY((float) y); }
+    public void set_focal_z(final double z) { client.setCameraFocalPointZ((float) z); }
 
     // Shake
     public boolean shake_disabled() { return client.isCameraShakeDisabled(); }
